pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neo-assistant"
version = "0.1.0"
description = "NEO: Neural Executive Operator - Multi-paradigm AI assistant"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "NEO Contributors"}
]
license = {text = "Apache-2.0"}
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Framework :: FastAPI",
  "License :: OSI Approved :: Apache Software License"
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-multipart>=0.0.9",
  "orjson>=3.10.0",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.0",
  "passlib[bcrypt]",
  "pyjwt>=2.8.0",
  "cryptography>=42.0.0",
  "structlog>=24.1.0",
  "tenacity>=8.2.0",
  "apscheduler>=3.10.4",
  "aiokafka>=0.10.0",
  "prometheus-client>=0.20.0",
  "websockets>=12.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "psutil>=5.9.0"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "ruff", "mypy", "black", "isort"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
